feat: add manual listing tool
